+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Fixed survey error when all options for a checkbox question were unselected
 - Fixed incorrect answer option migration mapping for the my-sleep-apnea survey
 - Hotkeys disabled for locked questions
+- Topics with slug `new` cannot be created.
 
 ## 5.0.0 (March 4, 2015)